// web/src/api/schema.d.ts
export interface components {
  schemas: {
// ...
    HourlyBucket: {
      /** Distinct Addresses */
      distinct_addresses: number;
      /**
       * Hour Utc
       * Format: date-time
       */
      hour_utc: string;
      /** Observations */
      observations: number;
      /** Scanner Id */
      scanner_id: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}